--- src/main/java/subway/ui/LineController.java
@@ -1,74 +1,0 @@
-package subway.ui;
-
-import java.net.URI;
-import java.sql.SQLException;
-import java.util.List;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-import subway.application.LineService;
-import subway.application.SectionService;
-import subway.application.StationService;
-import subway.dto.LineRequest;
-import subway.dto.LineResponse;
-import subway.dto.StationResponse;
-import subway.entity.Section;
-
-@RestController
-@RequestMapping("/lines")
-public class LineController {
-
-    private final LineService lineService;
-    private final StationService stationService;
-    private final SectionService sectionService;
-
-    public LineController(final LineService lineService, final StationService stationService,
-                          final SectionService sectionService) {
-        this.lineService = lineService;
-        this.stationService = stationService;
-        this.sectionService = sectionService;
-    }
-
-    @PostMapping
-    public ResponseEntity<Void> createLine(@RequestBody LineRequest lineRequest) {
-        final Long id = lineService.saveLine(lineRequest);
-        return ResponseEntity.created(URI.create("/lines/" + id)).build();
-    }
-
-    @GetMapping
-    public ResponseEntity<List<LineResponse>> findAllLines() {
-        return ResponseEntity.ok(lineService.findLineResponses());
-    }
-
-    @GetMapping("/{id}")
-    public ResponseEntity<LineResponse> findLineById(@PathVariable Long id) {
-        return ResponseEntity.ok(lineService.findLineResponseById(id));
-    }
-
-    @PutMapping("/{id}")
-    public ResponseEntity<Void> updateLine(@PathVariable Long id, @RequestBody LineRequest lineUpdateRequest) {
-        lineService.updateLine(id, lineUpdateRequest);
-        return ResponseEntity.ok().build();
-    }
-
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteLine(@PathVariable Long id) {
-        lineService.deleteLineById(id);
-        return ResponseEntity.noContent().build();
-    }
-
-    @GetMapping("/{id}/sections")
-    public ResponseEntity<List<StationResponse>> findAllStationOrderBySection(){
-        return ResponseEntity.ok(stationService.findAllStationOrderBySection());
-    }
-
-
-
-}
