--- src/main/java/subway/ui/StationController.java
@@ -1,51 +1,0 @@
-package subway.ui;
-
-import java.util.List;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-import subway.application.StationService;
-import subway.dto.StationDeleteRequest;
-import subway.dto.StationResponse;
-import subway.dto.StationSaveRequest;
-
-@RestController
-@RequestMapping("/stations")
-public class StationController {
-    private final StationService stationService;
-
-    public StationController(StationService stationService) {
-        this.stationService = stationService;
-    }
-
-    @PostMapping
-    public ResponseEntity<Void> createStation(@RequestBody StationSaveRequest stationSaveRequest) {
-        stationService.saveStation(stationSaveRequest);
-        return ResponseEntity.status(HttpStatus.CREATED).build();
-    }
-
-    @GetMapping
-    public ResponseEntity<List<StationResponse>> showStations() {
-        return ResponseEntity.ok().body(stationService.findAllStationResponses());
-    }
-
-    @GetMapping("/{id}")
-    public ResponseEntity<StationResponse> showStation(@PathVariable Long id) {
-        return ResponseEntity.ok().body(stationService.findStationResponseById(id));
-    }
-
-    @DeleteMapping()
-    public ResponseEntity<Void> deleteStation(@RequestBody StationDeleteRequest request) {
-        stationService.deleteStationById(request);
-        return ResponseEntity.noContent().build();
-    }
-
-}
